rename utils/ to util/, add edge injection comments
