+++ www/src/pages/llms-full.txt.ts
@@ -1,6 +1,6 @@
 import { getCollection } from "astro:content";
 import type { APIRoute } from "astro";
-import { cleanMarkdown } from "../utils/markdown";
+import { cleanMarkdown } from "../util/markdown";
 
 export const GET: APIRoute = async () => {
   const docs = await getCollection("docs");
